+++ Libft/ft_substr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jtomala <jtomala@student.42wolfsburg.de>   +#+  +:+       +#+        */
+/*   By: jtomala <jtomala@students.42wolfsburg.de>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/24 12:14:07 by jtomala           #+#    #+#             */
-/*   Updated: 2021/12/02 10:26:04 by jtomala          ###   ########.fr       */
+/*   Updated: 2022/04/15 08:32:12 by jtomala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	str = (char *)malloc(sizeof(char) * len + 1);
 	if (str == NULL)
 		return (NULL);
-	while (i < len && start <= ft_strlen(s))
+	while (i < len && start <= (unsigned int) ft_strlen(s))
 	{
 		str[i] = s[start + i];
 		i++;
